requirements.txt
streamlit>=1.33
pandas>=2.1
numpy>=1.26
plotly>=5.18
openpyxl>=3.1
pdfminer.six>=20221105
PyPDF2>=3.0
scipy>=1.11
scikit-learn>=1.3
sentence-transformers>=2.6
